+++ app/src/main/java/com/theandroiddev/mywins/core/injection/module/DataSourceModule.kt
@@ -1,9 +1,9 @@
 package com.theandroiddev.mywins.core.injection.module
 
-import com.theandroiddev.mywins.data.repository.SuccessImagesLocalDataSource
-import com.theandroiddev.mywins.data.source.SuccessImagesLocalDataSourceImpl
-import com.theandroiddev.mywins.data.repository.SuccessesLocalDataSource
-import com.theandroiddev.mywins.data.source.SuccessesLocalDataSourceImpl
+import com.theandroiddev.mywins.data.successes.repository.SuccessImagesLocalDataSource
+import com.theandroiddev.mywins.data.successes.source.SuccessImagesLocalDataSourceImpl
+import com.theandroiddev.mywins.data.successes.repository.SuccessesLocalDataSource
+import com.theandroiddev.mywins.data.successes.source.SuccessesLocalDataSourceImpl
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
